Reorganize Feed components, break FeedGroups out into its own folder
