+++ src/v2/components/FeedGroups/components/FeedGroupObjects/fragments/object.ts
