+++ src/v2/components/FeedGroups/components/FeedGroupObjects/fragments/objects.ts
@@ -1,6 +1,6 @@
 import gql from 'graphql-tag'
 
-import feedGroupObjectFragment from 'v2/components/Feed/components/FeedGroupObjects/fragments/object'
+import feedGroupObjectFragment from 'v2/components/FeedGroups/components/FeedGroupObjects/fragments/object'
 
 export default gql`
   fragment FeedGroupObjects on DeedGroup {
