+++ src/v2/components/FeedGroups/components/FeedGroupSentence/components/FeedObjectLink/index.js
