+++ src/v2/components/FeedGroups/components/FeedGroupSentence/fragments/object.ts
