+++ src/v2/components/FeedGroups/components/FeedGroupSentence/fragments/sentence.ts
@@ -1,6 +1,6 @@
 import gql from 'graphql-tag'
 
-import feedObjectFragment from 'v2/components/Feed/components/FeedGroupSentence/fragments/object'
+import feedObjectFragment from 'v2/components/FeedGroups/components/FeedGroupSentence/fragments/object'
 
 export default gql`
   fragment FeedGroupSentence on DeedGroup {
